requirements.txt
requests
pandas
psycopg2
statsmodels
tensorflow
scikit-learn
flask
supabase
python-dotenv